// src/web/src/api/types.gen.ts
export interface components {
    schemas: {
        Category: {
            /** Format: int32 */
            id?: number | string;
            name?: null | string;
        };
        FSharpOptionOfstring: string;
        GroupPhotosRequest: {
            sourceName?: null | string;
            targetName?: null | string;
        };
        GroupTreeNode: {
            groupId?: null | string;
            parentGroupId?: components["schemas"]["FSharpOptionOfstring"];
            photos?: null | string[];
        };
        MeResponse: {
            email: null | string;
        };
        MovePhotoToGroupRequest: {
            photoName?: null | string;
            targetGroupId?: null | string;
        };
        PhotoCount: {
            /** Format: date */
            date: string;
            /** Format: int32 */
            count: number | string;
        };
        PhotoInfo: {
            name: null | string;
            url: null | string;
            thumbnailUrl: null | string;
            /** Format: date-time */
            lastModified: string;
        };
        PhotoPage: {
            items: null | components["schemas"]["PhotoInfo"][];
            nextBefore: components["schemas"]["FSharpOptionOfstring"];
        };
        Suggestion: {
            groupId: null | string;
            reason: null | string;
        };
        VocabGroupDto: {
            id: null | string;
            name: null | string;
            subgroups: null | components["schemas"]["VocabSubgroupDto"][];
            ungroupedPhotos: null | components["schemas"]["VocabPhotoDto"][];
        };
        VocabMoveResponse: {
            vocabId: null | string;
        };
        VocabPhotoDto: {
            name: null | string;
            url: null | string;
            thumbnailUrl: null | string;
            lastModified: null | string;
        };
        VocabSubgroupDto: {
            id: null | string;
            photos: null | components["schemas"]["VocabPhotoDto"][];
        };
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}